\name{autoplot}
\alias{autoplot}
\title{Autoplot generic}
\usage{
autoplot(object, ...)
}
\arguments{
  \item{object}{An object to plot.}
  \item{...}{Additional arguments.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot \code{hrv_preprocessed} and \code{hrv_result} objects using the
\code{ggplot2} \code{autoplot()} generic.
}
